Update goldstar.cpp
about cmv4a , even if the erpoms 5-6-7 support tetris stealth game (GFX only) , the cmv4a is a regular cherry master 89 version . 
Very often you can meat roms with tetris game on it but the main program not support that. Also all the tetris cherry master or with pacman front games the main program should be on 27c512 eprom . Not enough space for front game for 32k 27c256 eproms.
Also the GAME: cmv801 has tetris into the main program (prg512 000065536 main cpu) but is a modified version and boot direct on cherry side.All the tetris games needs the same combination for enter on cherry side.